Rename classes for consistency and move around packages
