+++ krystex/src/main/java/com/flipkart/krystal/krystex/caching/RequestLevelCache.java
@@ -1,6 +1,6 @@
 package com.flipkart.krystal.krystex.caching;
 
-import static com.flipkart.krystal.utils.Futures.linkFutures;
+import static com.flipkart.krystal.concurrent.Futures.linkFutures;
 import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static java.util.concurrent.CompletableFuture.allOf;
