+++ vajram/vajram-samples/src/test/java/com/flipkart/krystal/vajram/samples/calculator/DoubleMinusOneTest.java
@@ -1,8 +1,12 @@
 package com.flipkart.krystal.vajram.samples.calculator;
 
 import static com.flipkart.krystal.vajram.VajramID.ofVajram;
-import static org.junit.jupiter.api.Assertions.*;
 
+import com.flipkart.krystal.concurrent.SingleThreadExecutor;
+import com.flipkart.krystal.concurrent.SingleThreadExecutorsPool;
+import com.flipkart.krystal.krystex.kryon.KryonExecutorConfig;
+import com.flipkart.krystal.pooling.Lease;
+import com.flipkart.krystal.pooling.LeaseUnavailableException;
 import com.flipkart.krystal.vajramexecutor.krystex.KrystexVajramExecutor;
 import com.flipkart.krystal.vajramexecutor.krystex.KrystexVajramExecutorConfig;
 import com.flipkart.krystal.vajramexecutor.krystex.VajramKryonGraph;
@@ -10,19 +14,45 @@ import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class DoubleMinusOneTest {
+  private static SingleThreadExecutorsPool EXEC_POOL;
+
+  @BeforeAll
+  static void beforeAll() {
+    EXEC_POOL =
+        new SingleThreadExecutorsPool(
+            "RequestLevelCacheTest", Runtime.getRuntime().availableProcessors());
+  }
 
   private final VajramKryonGraph graph =
       VajramKryonGraph.builder().loadFromPackage(DoubleMinusOne.class.getPackageName()).build();
+  private Lease<SingleThreadExecutor> executorLease;
+
+  @BeforeEach
+  void setUp() throws LeaseUnavailableException {
+    this.executorLease = EXEC_POOL.lease();
+  }
+
+  @AfterEach
+  void tearDown() {
+    executorLease.close();
+  }
 
   @Test
   void doubleMinusOne_success() {
     CompletableFuture<Integer> future;
     try (KrystexVajramExecutor krystexVajramExecutor =
         graph.createExecutor(
-            KrystexVajramExecutorConfig.builder().requestId("doubleMinusOne").build())) {
+            KrystexVajramExecutorConfig.builder()
+                .requestId("doubleMinusOne")
+                .kryonExecutorConfigBuilder(
+                    KryonExecutorConfig.builder().singleThreadExecutor(executorLease.get()))
+                .build())) {
       future =
           krystexVajramExecutor.execute(
               ofVajram(DoubleMinusOne.class),
