+++ vajram/vajram-samples/src/test/java/com/flipkart/krystal/vajram/samples_errors/TooManyQualifiersTest.java
@@ -6,6 +6,11 @@ import static com.google.inject.name.Names.named;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.flipkart.krystal.concurrent.SingleThreadExecutor;
+import com.flipkart.krystal.concurrent.SingleThreadExecutorsPool;
+import com.flipkart.krystal.krystex.kryon.KryonExecutorConfig;
+import com.flipkart.krystal.pooling.Lease;
+import com.flipkart.krystal.pooling.LeaseUnavailableException;
 import com.flipkart.krystal.vajram.MandatoryFacetsMissingException;
 import com.flipkart.krystal.vajram.guice.VajramGuiceInjector;
 import com.flipkart.krystal.vajramexecutor.krystex.KrystexVajramExecutor;
@@ -15,17 +20,34 @@ import com.flipkart.krystal.vajramexecutor.krystex.VajramKryonGraph.Builder;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import org.checkerframework.checker.nullness.qual.Nullable;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class TooManyQualifiersTest {
+  private static SingleThreadExecutorsPool EXEC_POOL;
+
+  @BeforeAll
+  static void beforeAll() {
+    EXEC_POOL =
+        new SingleThreadExecutorsPool(
+            "RequestLevelCacheTest", Runtime.getRuntime().availableProcessors());
+  }
 
   private Builder graph;
+  private Lease<SingleThreadExecutor> executorLease;
 
   @BeforeEach
-  void setUp() {
+  void setUp() throws LeaseUnavailableException {
+    this.executorLease = EXEC_POOL.lease();
     graph =
         new VajramKryonGraph.Builder().loadFromPackage(TooManyQualifiers.class.getPackageName());
+  }
+
+  @AfterEach
+  void tearDown() {
+    executorLease.close();
   }
 
   @Test
@@ -48,9 +70,11 @@ class TooManyQualifiersTest {
                 + "'inject' of vajram 'TooManyQualifiers'. This is not allowed");
   }
 
-  private static KrystexVajramExecutor createExecutor(VajramKryonGraph vajramKryonGraph) {
+  private KrystexVajramExecutor createExecutor(VajramKryonGraph vajramKryonGraph) {
     return vajramKryonGraph.createExecutor(
         KrystexVajramExecutorConfig.builder()
+            .kryonExecutorConfigBuilder(
+                KryonExecutorConfig.builder().singleThreadExecutor(executorLease.get()))
             .inputInjectionProvider(
                 new VajramGuiceInjector(
                     createInjector(
